Splitting DB and Pydantic models
