+++ app/database/base_table_model.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, TIMESTAMP, String, Integer
 
-from app.database import Base
+from app.database.database import Base
 
 class BaseTableModel(Base):
     __abstract__ = True
