+++ app/main.py
@@ -19,7 +19,7 @@ from app.services.user import UserService
 app = FastAPI()
 app.add_middleware(
     CORSMiddleware,
-    allow_origins=["http://localhost:4200"],  # Your Angular app's URL (adjust as needed)
+    allow_origins=["*"],  # Your Angular app's URL (adjust as needed)
     allow_credentials=True,
     allow_methods=["*"],  # Or specify: ["GET", "POST", "PUT", "DELETE", etc.]
     allow_headers=["*"],  # Or specify required headers
