+++ app/routers/__init__.py
@@ -1,0 +1,6 @@
+from app.routers.enum_type import EnumTypeRouter
+from app.routers.health import HealthRouter
+from app.routers.login import LoginRouter
+from app.routers.member import MemberRouter
+from app.routers.preaching_point import PreachingPointRouter
+from app.routers.user import UserRouter
