+++ app/routers/preaching_point.py
@@ -3,7 +3,7 @@ from typing import Optional, List
 from fastapi import Depends, APIRouter
 from sqlalchemy.orm import Session
 
-from app.database import get_db
+from app.database.database import get_db
 from app.models.preaching_point import PreachingPointInformation
 from app.services.auth import AuthService
 from app.services.preaching_point import PreachingPointService
